Convert encoded MAE tokens to correct type (#1094)
* Convert encoded MAE tokens to correct type

This solves an issue when training MAE with half-precision.
In that case, we want to widen the type of the encoded tokens
to the same dtype as the mask token.

+++ examples/pytorch_lightning_distributed/mae.py
@@ -47,7 +47,7 @@ class MAE(pl.LightningModule):
         x_masked = utils.repeat_token(
             self.mask_token, (batch_size, self.sequence_length)
         )
-        x_masked = utils.set_at_index(x_masked, idx_keep, x_decode)
+        x_masked = utils.set_at_index(x_masked, idx_keep, x_decode.type_as(x_masked))
 
         # decoder forward pass
         x_decoded = self.decoder.decode(x_masked)
